+++ src/Stratis.Bitcoin.Features.SmartContracts/ISmartContractCoinviewRule.cs
