+++ src/Stratis.Bitcoin.Features.SmartContracts/PoA/SmartContractPoABlockDefinition.cs
