+++ src/Stratis.Bitcoin.Features.SmartContracts/PoA/SmartContractPoAConsensusFactory.cs
