+++ src/Stratis.Bitcoin.Features.SmartContracts/PoA/SmartContractPoARuleEngine.cs
