+++ src/Stratis.Bitcoin.Features.SmartContracts/PoA/SmartContractPoARuleRegistration.cs
