+++ src/Stratis.Bitcoin.Features.SmartContracts/PoS/Rules/SmartContractPosBlockSignatureRule.cs
