+++ src/Stratis.Bitcoin.Features.SmartContracts/PoS/Rules/SmartContractPosCoinviewRule.cs
