+++ src/Stratis.Bitcoin.Features.SmartContracts/PoS/SmartContractBlockProvider.cs
