+++ src/Stratis.Bitcoin.Features.SmartContracts/PoS/SmartContractPosBlock.cs
