+++ src/Stratis.Bitcoin.Features.SmartContracts/PoS/SmartContractPosBlockHeader.cs
