+++ src/Stratis.Bitcoin.Features.SmartContracts/PoS/SmartContractPosConsensusFactory.cs
