+++ src/Stratis.Bitcoin.Features.SmartContracts/PoS/SmartContractPosConsensusRuleEngine.cs
