+++ src/Stratis.Bitcoin.Features.SmartContracts/PoS/SmartContractPosPowBlockDefinition.cs
