+++ src/Stratis.Bitcoin.Features.SmartContracts/PoS/SmartContractPosRuleRegistration.cs
