+++ src/Stratis.Bitcoin.Features.SmartContracts/PoS/SmartContractPosTransaction.cs
