+++ src/Stratis.Bitcoin.Features.SmartContracts/PoW/Rules/SmartContractPowCoinviewRule.cs
