+++ src/Stratis.Bitcoin.Features.SmartContracts/PoW/SmartContractBlockDefinition.cs
