+++ src/Stratis.Bitcoin.Features.SmartContracts/PoW/SmartContractPowBlockHeader.cs
