+++ src/Stratis.Bitcoin.Features.SmartContracts/PoW/SmartContractPowConsensusFactory.cs
