+++ src/Stratis.Bitcoin.Features.SmartContracts/PoW/SmartContractPowConsensusRuleEngine.cs
