+++ src/Stratis.Bitcoin.Features.SmartContracts/PoW/SmartContractPowRuleRegistration.cs
