+++ src/Stratis.Bitcoin.Features.SmartContracts/Rules/AllowedScriptTypeRule.cs
