+++ src/Stratis.Bitcoin.Features.SmartContracts/Rules/CanGetSenderRule.cs
