+++ src/Stratis.Bitcoin.Features.SmartContracts/Rules/MempoolOpSpendRule.cs
