+++ src/Stratis.Bitcoin.Features.SmartContracts/Rules/OpSpendRule.cs
