+++ src/Stratis.Bitcoin.Features.SmartContracts/Rules/P2PKHNotContractRule.cs
