+++ src/Stratis.Bitcoin.Features.SmartContracts/Rules/SmartContractCoinviewRule.cs
