+++ src/Stratis.Bitcoin.Features.SmartContracts/Rules/TxOutSmartContractExecRule.cs
