+++ src/Stratis.Bitcoin.Features.SmartContracts/SmartContractBlockHeader.cs
