+++ src/Stratis.Bitcoin.Features.SmartContracts/SmartContractFeature.cs
@@ -125,173 +125,6 @@ namespace Stratis.Bitcoin.Features.SmartContracts
         }
 
         /// <summary>
-        /// Configures the node with the smart contract proof of work consensus model.
-        /// </summary>
-        public static IFullNodeBuilder UseSmartContractPowConsensus(this IFullNodeBuilder fullNodeBuilder)
-        {
-            LoggingConfiguration.RegisterFeatureNamespace<ConsensusFeature>("consensus");
-
-            fullNodeBuilder.ConfigureFeature(features =>
-            {
-                features
-                .AddFeature<ConsensusFeature>()
-                .DependOn<SmartContractFeature>()
-                .FeatureServices(services =>
-                {
-                    services.AddSingleton<ConsensusOptions, ConsensusOptions>();
-                    services.AddSingleton<DBreezeCoinView>();
-                    services.AddSingleton<ICoinView, CachedCoinView>();
-                    services.AddSingleton<ConsensusController>();
-                    services.AddSingleton<IConsensusRuleEngine, SmartContractPowConsensusRuleEngine>();
-
-                    new SmartContractPowRuleRegistration(fullNodeBuilder.Network).RegisterRules(fullNodeBuilder.Network.Consensus);
-                });
-            });
-
-            return fullNodeBuilder;
-        }
-
-        /// <summary>
-        /// Configures the node with the smart contract proof of authority consensus model.
-        /// </summary>
-        public static IFullNodeBuilder UseSmartContractPoAConsensus(this IFullNodeBuilder fullNodeBuilder)
-        {
-            LoggingConfiguration.RegisterFeatureNamespace<ConsensusFeature>("consensus");
-
-            fullNodeBuilder.ConfigureFeature(features =>
-            {
-                features
-                    .AddFeature<ConsensusFeature>()
-                    .DependOn<SmartContractFeature>()
-                    .FeatureServices(services =>
-                    {
-                        services.AddSingleton<DBreezeCoinView>();
-                        services.AddSingleton<ICoinView, CachedCoinView>();
-                        services.AddSingleton<ConsensusController>();
-                        services.AddSingleton<IConsensusRuleEngine, SmartContractPoARuleEngine>();
-                        
-                        new SmartContractPoARuleRegistration(fullNodeBuilder.Network).RegisterRules(fullNodeBuilder.Network.Consensus);
-                    });
-            });
-
-            return fullNodeBuilder;
-        }
-
-
-        /// <summary>
-        /// Configures the node with the smart contract proof of stake consensus model.
-        /// </summary>
-        public static IFullNodeBuilder UseSmartContractPosConsensus(this IFullNodeBuilder fullNodeBuilder)
-        {
-            LoggingConfiguration.RegisterFeatureNamespace<ConsensusFeature>("consensus");
-
-            fullNodeBuilder.ConfigureFeature(features =>
-            {
-                features
-                    .AddFeature<ConsensusFeature>()
-                    .FeatureServices(services =>
-                    {
-                        services.AddSingleton<DBreezeCoinView>();
-                        services.AddSingleton<ICoinView, CachedCoinView>();
-                        services.AddSingleton<StakeChainStore>().AddSingleton<IStakeChain, StakeChainStore>(provider => provider.GetService<StakeChainStore>());
-                        services.AddSingleton<IStakeValidator, StakeValidator>();
-                        services.AddSingleton<ConsensusController>();
-                        services.AddSingleton<IConsensusRuleEngine, SmartContractPosConsensusRuleEngine>();
-
-                        new SmartContractPosRuleRegistration().RegisterRules(fullNodeBuilder.Network.Consensus);
-                    });
-            });
-
-            return fullNodeBuilder;
-        }
-
-        /// <summary>
-        /// Adds mining to the smart contract node.
-        /// <para>We inject <see cref="IPowMining"/> with a smart contract block provider and definition.</para>
-        /// </summary>
-        public static IFullNodeBuilder UseSmartContractPowMining(this IFullNodeBuilder fullNodeBuilder)
-        {
-            LoggingConfiguration.RegisterFeatureNamespace<MiningFeature>("mining");
-
-            fullNodeBuilder.ConfigureFeature(features =>
-            {
-                features
-                    .AddFeature<MiningFeature>()
-                    .DependOn<MempoolFeature>()
-                    .DependOn<RPCFeature>()
-                    .DependOn<SmartContractWalletFeature>()
-                    .FeatureServices(services =>
-                    {
-                        services.AddSingleton<IPowMining, PowMining>();
-                        services.AddSingleton<IBlockProvider, SmartContractBlockProvider>();
-                        services.AddSingleton<BlockDefinition, SmartContractBlockDefinition>();
-                        services.AddSingleton<IBlockBufferGenerator, BlockBufferGenerator>();
-                        services.AddSingleton<MiningController>();
-                        services.AddSingleton<MiningRpcController>();
-                        services.AddSingleton<MinerSettings>();
-                    });
-            });
-
-            return fullNodeBuilder;
-        }
-
-        /// <summary>
-        /// Adds mining to the smart contract node.
-        /// <para>We inject <see cref="IPowMining"/> with a smart contract block provider and definition.</para>
-        /// </summary>
-        public static IFullNodeBuilder UseSmartContractPosPowMining(this IFullNodeBuilder fullNodeBuilder)
-        {
-            LoggingConfiguration.RegisterFeatureNamespace<MiningFeature>("mining");
-
-            fullNodeBuilder.ConfigureFeature(features =>
-            {
-                features
-                    .AddFeature<MiningFeature>()
-                    .DependOn<MempoolFeature>()
-                    .DependOn<RPCFeature>()
-                    .DependOn<SmartContractWalletFeature>()
-                    .FeatureServices(services =>
-                    {
-                        services.AddSingleton<IPowMining, PowMining>();
-                        services.AddSingleton<IBlockProvider, SmartContractBlockProvider>();
-                        services.AddSingleton<BlockDefinition, SmartContractBlockDefinition>();
-                        services.AddSingleton<BlockDefinition, SmartContractPosPowBlockDefinition>();
-                        services.AddSingleton<IBlockBufferGenerator, BlockBufferGenerator>();
-                        services.AddSingleton<MiningRpcController>();
-                        services.AddSingleton<MiningController>();
-                        services.AddSingleton<StakingController>();
-                        services.AddSingleton<StakingRpcController>();
-                        services.AddSingleton<MinerSettings>();
-                    });
-            });
-
-            return fullNodeBuilder;
-        }
-
-        /// <summary>
-        /// Adds mining to the smart contract node when on a proof-of-authority network.
-        /// </summary>
-        public static IFullNodeBuilder UseSmartContractPoAMining(this IFullNodeBuilder fullNodeBuilder)
-        {
-            fullNodeBuilder.ConfigureFeature(features =>
-            {
-                features
-                    .AddFeature<PoAFeature>()
-                    .FeatureServices(services =>
-                    {
-                        services.AddSingleton<FederationManager>();
-                        services.AddSingleton<PoABlockHeaderValidator>();
-                        services.AddSingleton<IPoAMiner, PoAMiner>();
-                        services.AddSingleton<SlotsManager>();
-                        services.AddSingleton<BlockDefinition, SmartContractPoABlockDefinition>();
-                        services.AddSingleton<IBlockBufferGenerator, BlockBufferGenerator>();
-                    });
-            });
-
-            return fullNodeBuilder;
-        }
-
-        /// <summary>
         /// This node will be configured with the reflection contract executor.
         /// <para>
         /// Should we require another executor, we will need to create a separate daemon and network.
