+++ src/Stratis.Bitcoin.Features.SmartContracts/SmartContractScriptAddressReader.cs
